platform/atrace-capture/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atrace-capture"
version = "1.0.0"
description = "Capture engine layer for TraceMind — CaptureEngine ABC, config schema, presets"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "atrace-device",
    "atrace-provision",
]

# Monorepo: `uv` resolves sibling packages as path deps (plain `pip install -e`
# from PyPI still needs peers installed separately — use ./dev-setup.sh).
[tool.uv.sources]
atrace-device = { path = "../atrace-device", editable = true }
atrace-provision = { path = "../atrace-provision", editable = true }

[tool.setuptools.packages.find]
where = ["."]
include = ["atrace_capture*"]

[tool.setuptools.package-data]
"atrace_capture.config" = ["presets/*.yaml", "perfetto/*.txtpb", "perfetto/README.md"]
